+++ src/dom/domnode.ts
@@ -1,0 +1,7 @@
+export class DOMNode {
+	readonly nodeName: string;
+
+	constructor(nodeName: string) {
+		this.nodeName = nodeName;
+	}
+}