+++ src/dom/index.ts
@@ -1,4 +1,5 @@
 export { Attribute } from './attribute';
 export { HtmlElement, NodeClosed } from './htmlelement';
+export { DOMNode } from './domnode';
 export { DOMTokenList } from './domtokenlist';
 export { DOMTree } from './domtree';