+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Breaking changes
 
-- `DOMNode` has been renamed `HtmlElement`.
+- `DOMNode` has been renamed `HtmlElement` and there is instead a new base class
+  `DOMnode` which `HtmlElement` extends from. Rules using `DOMNode` need to be
+  changed to use `HtmlElement`.
 
 ### Features